requirements.txt
pymongo==4.6.0
pandas==2.1.4
python-dotenv==1.0.0
confluent-kafka==2.3.0
pymongo==4.6.1